+++ HHSNewsVC.m
@@ -173,7 +173,7 @@
 
 - (void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    [self sendToDetailPager:indexPath.row parentViewController:self];
+    [self sendToDetailPager:(int)indexPath.row parentViewController:self];
 }
 
 - (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
